server/requirements.txt
django-filter
drf_yasg